requirements.txt
flask>=2.3.0
pyserial>=3.5
gunicorn>=21.0.0
reportlab>=4.0.0
